WordPress PHP file
This is the code that I used to wire JSON data generated from a custom WP admin interface (that I created) to the actual WP site. The site is a drag & drop CMS (that I only styled, I didn't create) where the modules and images that are available are managed from the custom WP admin. I'd never done anything like this before and had a great time crash-coursing my way through it.